synthetic_medical_imaging/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0

# LLM providers
anthropic>=0.8.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
